+++ src/users/users.service.ts
@@ -6,6 +6,8 @@ import { PaginatedResult } from 'src/common/interfaces/paginated-result.interfac
 import { MutateUserResponseDto } from './dto/mutate-users-response.dto';
 import { FileService } from 'src/file/file.service';
 import { DropdownUsersResponseDto } from './dto/dropdown-users.dto';
+import { CreateUserDto } from './dto/create-users.dto';
+import { UpdateUserDto } from './dto/update-users.dto';
 
 @Injectable()
 export class UsersService {
@@ -15,8 +17,9 @@ export class UsersService {
     ){}
 
     async create(
-        data: Prisma.UserCreateInput
+        data: CreateUserDto
     ): Promise<MutateUserResponseDto> {
+        const { hospitalId, rateListFileName, ...rest } = data
         const existingUser = await this.prisma.user.findUnique({
             where: { email: data.email.toLowerCase() },
         });
@@ -25,9 +28,11 @@ export class UsersService {
         const hashedPassword = await bcrypt.hash(data.password, 10);
         return await this.prisma.user.create({
             data: {
-                ...data,
+                ...rest,
                 email: data.email.toLowerCase(),
                 password: hashedPassword,
+                ...(hospitalId && { hospital: { connect: { id: hospitalId }}}),
+                ...((data.role===Role.HOSPITAL && rateListFileName) && { rateListFileName })
             },
             select: {
                 id: true,
@@ -37,6 +42,7 @@ export class UsersService {
                 hospitalName: true,
                 rateListFileName: true,
                 rateListUrl: true,
+                hospital: { select: { id: true, name: true }},
                 role: true,
             }
         });
@@ -81,6 +87,7 @@ export class UsersService {
                     email: true,
                     address: true,
                     hospitalName: true,
+                    hospital: { select: { id: true, name: true }},
                     role: true,
                 }
             })
@@ -103,6 +110,7 @@ export class UsersService {
                 profileUrl: true,
                 rateListFileName: true,
                 rateListUrl: true,
+                hospital: { select: { id: true, name: true }},
                 role: true,
             }
         })
@@ -118,12 +126,22 @@ export class UsersService {
 
     async update(params: {
         where: Prisma.UserWhereUniqueInput,
-        data: Prisma.UserUpdateInput
+        data: UpdateUserDto
     }): Promise<MutateUserResponseDto> {
         const { where, data } = params;
-        const { email, password, ...rest } = data
+        const { email, password, hospitalId, ...rest } = data
+         const isSuperAdminUser = await this.prisma.user.findFirst({
+            where: { 
+                id: where.id,
+                role: Role.SUPER_ADMIN
+            }
+        })
+        if(isSuperAdminUser) throw new BadRequestException("SUPERADMIN role can't be modified!")
         return this.prisma.user.update({ 
-            data: { ...rest }, 
+            data: { 
+                ...rest,
+                ...(hospitalId && { hospital: { connect: { id: hospitalId }}})
+            }, 
             where,
             select: {
                 id: true,
@@ -133,6 +151,7 @@ export class UsersService {
                 hospitalName: true,
                 rateListFileName: true,
                 rateListUrl: true,
+                hospital: { select: { id: true, name: true }},
                 role: true,
             }
         })
@@ -158,20 +177,33 @@ export class UsersService {
     }
 
     async remove(
-        where: Prisma.UserWhereUniqueInput
+        whereUniqueInput: Prisma.UserWhereUniqueInput
     ): Promise<MutateUserResponseDto>{
-        return this.prisma.user.delete({
-            where,
+        const isUserExists = await this.prisma.user.findFirst({
+            where: whereUniqueInput,
+            select: { id: true, role: true }
+        })
+        if(isUserExists.role===Role.SUPER_ADMIN) throw new BadRequestException("SUPERADMIN role can't be modified!")
+
+        const result =  await this.prisma.user.delete({
+            where: whereUniqueInput,
             select: {
                 id: true,
                 name: true,
                 email: true,
-                address: true,
-                hospitalName: true,
+                profileFileName: true,
                 rateListFileName: true,
-                rateListUrl: true,
                 role: true,
             }
         })
+        const { profileFileName, rateListFileName} = result
+        try {
+            if(profileFileName) await this.fileService.deleteFile(profileFileName)
+            if(rateListFileName) await this.fileService.deleteFile(rateListFileName)
+        } catch (error) {
+            console.error("USER_DELETE_SERVICE_S3: ", error)
+        }
+
+        return result
     }
 }
